requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
exceptiongroup==1.2.0
frozenlist==1.4.1
gpt4all==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jellyfish==1.0.3
llm==0.12
llm-gpt4all==0.2
multidict==6.0.4
numpy==1.26.2
openai==0.28.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-ulid==2.2.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.36
tabulate==0.9.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
yarl==1.9.4
